requirement.txt
fastapi
psycopg[binary]
python-dotenv
sqlalchemy
pyjwt
pwdlib[argon2]
pydantic[email]
python-multipart